feat: migrate to App Router Api refactor: pocemonListWrapper
