+++ app/page.tsx
@@ -1,16 +1,23 @@
+'use client';
 import React from 'react';
 import SimulateErrorButton from '../src/components/ErrorBoundary/SimulateErrorButton';
 import Flyout from '../src/components/Flyout/Flyout';
-import PokemonListWrapper from '../src/components/PokemonList/PokemonListWrapper';
-import Layout from '../src/components/Layout';
+import dynamic from 'next/dynamic';
+
+const PokemonListWrapper = dynamic(
+  () => import('../src/components/PokemonList/PokemonListWrapper'),
+  {
+    ssr: false,
+  }
+);
 
 const HomePage: React.FC = () => {
   return (
-    <Layout>
+    <>
       <SimulateErrorButton onClick={() => console.error('Simulated Error')} />
       <PokemonListWrapper />
       <Flyout />
-    </Layout>
+    </>
   );
 };
