+++ next.config.js
@@ -1,6 +1,12 @@
 module.exports = {
-    images: {
-      domains: ['raw.githubusercontent.com'],
-    },
-  };
-  
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'raw.githubusercontent.com',
+        port: '',
+        pathname: '/**',
+      },
+    ],
+  },
+};
