+++ package.json
@@ -52,6 +52,7 @@
     "@typescript-eslint/parser": "^7.15.0",
     "autoprefixer": "^10.4.20",
     "eslint": "^8.57.0",
+    "eslint-config-next": "^14.2.5",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-next": "^0.0.0",
     "eslint-plugin-prettier": "^5.1.3",
